@@ -11,21 +11,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^14.2.9",
+    "@angular-devkit/build-angular": "^14.2.7",
     "@angular-devkit/build-ng-packagr": "^0.1002.0",
-    "@angular/animations": "^14.2.10",
+    "@angular/animations": "^14.2.7",
     "@angular/cdk": "^14.2.7",
     "@angular/cli": "^9.1.8",
-    "@angular/common": "^14.2.10",
-    "@angular/compiler": "^10.2.5",
-    "@angular/core": "^14.2.10",
-    "@angular/forms": "^14.2.10",
-    "@angular/http": "^7.2.15",
+    "@angular/common": "^14.2.7",
+    "@angular/compiler": "^14.2.7",
+    "@angular/core": "^14.2.7",
+    "@angular/forms": "^14.2.7",
     "@angular/material": "^14.2.7",
-    "@angular/platform-browser": "^14.2.10",
-    "@angular/platform-browser-dynamic": "^14.2.10",
-    "@angular/router": "^14.2.10",
-    "angular-google-charts": "^2.2.0",
+    "@angular/platform-browser": "^14.2.7",
+    "@angular/platform-browser-dynamic": "^14.2.7",
+    "@angular/router": "^14.2.7",
+    "angular-google-charts": "^2.2.3",
     "bootstrap": "4.3.1",
     "chart.js": "^2.9.4",
     "core-js": "^3.4.1",
@@ -34,25 +33,24 @@
     "ngx-bootstrap": "^5.2.0",
     "rxjs": "^6.5.4",
     "rxjs-compat": "^6.5.3",
-    "zone.js": "^0.10.2"
+    "zone.js": "^0.12.0"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^10.2.5",
-    "@angular/language-service": "^10.2.5",
-    "@types/jasmine": "^3.5.0",
+    "@angular/compiler-cli": "^14.2.7",
+    "@angular/language-service": "^14.2.7",
+    "@types/jasmine": "^3.8.0",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "~12.12.9",
-    "codelyzer": "^0.0.28",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~3.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.5.2",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.8"
+    "tslint": "~6.1.3",
+    "typescript": "~4.6.2"
   }
 }
