@@ -6,9 +6,9 @@ import {
   Output
 } from "@angular/core";
 import { AuthenticationService } from "../../auth/authentication.service";
-import { Router, ActivatedRoute, RouterEvent } from "@angular/router";
+import { Router, ActivatedRoute, RouterEvent, NavigationStart, NavigationEnd } from "@angular/router";
 import { DatePipe } from "@angular/common";
-import { DATE_FORMAT, DETAIL_DATE_FORMAT } from "../../app.constants";
+import { ACTIVE_PATH, DATE_FORMAT, DETAIL_DATE_FORMAT } from "../../app.constants";
 import { filter } from "rxjs/operators";
 import { ImportantEditComponent } from "src/app/important/important-edit/important-edit.component";
 import { Subscription } from "rxjs";
@@ -33,43 +33,30 @@ export class HeaderComponent implements OnInit {
   @Input()
   second: ImportantEditComponent;
 
-  @Output()
-  dataChangedEvent = new EventEmitter();
-
   constructor(
-    private authService: AuthenticationService,
+    public authService: AuthenticationService,
     private datepipe: DatePipe,
-    private router: Router,
-    private route: ActivatedRoute
+    private router: Router
   ) {
     this.router.events.pipe(
       filter(event => event instanceof RouterEvent)
     ).subscribe((event: RouterEvent) => {
-      if(event) {
-          // console.log("Event: " + event.url);
-          // console.log(event.url.includes("important"));
-          this.activeUrl = "active";
+      if(event instanceof NavigationStart || event instanceof NavigationEnd) {
+          let active = event.url.split("/")[1];
+          this.activeUrl = active;
+          sessionStorage.setItem(ACTIVE_PATH, active);
       }
     });
   }
 
   ngOnInit() {
+    console.log('ngOnInit');
     this.date = new Date();
-    this.username = this.authService.getAuthenticatedUser();
     this.day = this.date.getDate();
     this.month = this.date.getMonth() + 1;
     this.year = this.date.getFullYear();
     this.startDate = this.datepipe.transform(new Date(this.year, this.month - 1, this.day), DATE_FORMAT);
-    this.endDate = this.datepipe.transform(new Date(), DATE_FORMAT);
-    this.dataChangedEvent.emit('event');
+    this.endDate = this.datepipe.transform(new Date(), DATE_FORMAT);  
   }
 
-  toActivate(url: string) {
-    return this.activeUrl.includes(url) ? "active" : "";
-  }
-
-  loadData() {
-    console.log('event occured --- ');
-    this.dataChangedEvent.emit('event');
-  }
 }
