@@ -1,7 +1,7 @@
 import { Component, OnInit, Inject } from "@angular/core";
 import { Extraordinary } from "../extraordinary.model";
 import { Router } from "@angular/router";
-import { MAT_DIALOG_DATA } from "@angular/material";
+import { MAT_DIALOG_DATA } from "@angular/material/dialog";
 import { ExtraordinaryService } from "src/app/extraordinary/extraordinary.service";
 import { AuthenticationService } from "src/app/auth/authentication.service";
 import { ErrorService } from "src/app/services/data/error.service";
